Rework ametsuchi test:  - remove namespace  - move test fixture in ametshchi mocks  - make interface more clear (via using)
